Added more tables and entities for them

+++ dao/src/main/java/com/gmail/evanloafakahaitao/ItemDao.java
@@ -13,5 +13,5 @@ public interface ItemDao {
 
     Item findByVendorCode(Connection connection, Long vendorCode);
 
-    Item findById(Connection connection, int itemId);
+    Item findById(Connection connection, Long itemId);
 }
